fix: Assignment autograde
It seems that if we use existing task definition then nbgrader will run within
the same container which is not ideal, because ECS will try to run it on the same EC2 instance.
When there is no resources available on that instance then it will fail.
So I created new task definition and run nbgrader as completaly separate task.

Signed-off-by: Jan Kowalski <jankowalskipraca85@gmail.com>